requirements.txt
prettyconf
pytest
python-telegram-bot
emoji
requests
cheetah3
